+++ controllers/tests/pick_up.py
@@ -16,10 +16,10 @@ class PickUp(TestController):
 
     def init_scene(self, scene: str = None, layout: int = None, room: int = None) -> ActionStatus:
         # Add some objects to an empty room. Record their object IDs.
-        self.target_object_0 = self._add_object("jug05", position={"x": -0.408, "y": 0, "z": 0.428})
+        self.target_object_0 = self._add_object("jug05", position={"x": -0.408, "y": 0, "z": 0.428},)
         self.target_object_1 = self._add_object("jug05", position={"x": -1.76, "y": 0, "z": -1.08})
         self.box = self._add_object("basket_18inx18inx12iin", position={"x": 0.03, "y": 0, "z": -2.38},
-                                    scale={"x": 1, "y": 0.4, "z": 1})
+                                    scale={"x": 1, "y": 0.5, "z": 1})
 
         return super().init_scene()
 
@@ -51,7 +51,7 @@ if __name__ == "__main__":
 
     # Get a point above the box.
     box_top = m.state.object_transforms[m.box].position[:]
-    box_top[1] += m.objects_static[m.box].size[1] + 0.6
+    box_top[1] += m.objects_static[m.box].size[1] + 0.7
 
     # Drop each object.
     for arm, object_id in zip([Arm.left, Arm.right], [m.target_object_0, m.target_object_1]):